+++ src/NeuralNet/CostFunctions/CrossEntropy.php
@@ -23,16 +23,6 @@ use const Rubix\ML\EPSILON;
 class CrossEntropy implements ClassificationLoss
 {
     /**
-     * Return a tuple of the min and max output value for this function.
-     *
-     * @return float[]
-     */
-    public function range() : array
-    {
-        return [0.0, INF];
-    }
-
-    /**
      * Compute the loss score.
      *
      * @param \Tensor\Matrix $output