+++ src/NeuralNet/Layers/Dropout.php
@@ -61,7 +61,7 @@ class Dropout implements Hidden
     public function __construct(float $ratio = 0.5)
     {
         if ($ratio <= 0.0 or $ratio >= 1.0) {
-            throw new InvalidArgumentException('Ratio must be strictly'
+            throw new InvalidArgumentException('Ratio must be'
                 . " between 0 and 1, $ratio given.");
         }